First try at fail-safing the mail sending
Added try-catch logic to UserMailer to test with
Also made it such that the login and password of the email account are
no longer stored in plain text in the repository. If this works, need
to update the other environments and change the password so the account
is slightly more secure.